Add write flag

+++ src/python/util/crud.py
@@ -8,19 +8,26 @@ from .wrap import *
 
 # Create
 @decor_path
-def create_file(path):
-    pass
+def create_file(path, data):
+    if os.path.isfile(path):
+        raise NotImplementedError(f'File already exist: "{path}"')
+    if VARS['WRITE']:
+        print_info(f'Creating file: "{path}"')
+        bash_cmd(f'touch "{path}"')
+        bash_cmd(f'printf "{data}" > "{path}"')
 
 
 @decor_path
 def create_path(path):
-    if not os.path.isdir(path):
+    if VARS['WRITE'] and not os.path.isdir(path):
         bash_cmd(f'mkdir -p "{path}"')
 
 
 @decor_path_args(DOTFILES, HOME)
 def link_file(src, dst, desc=''):
     '''Link file'''
+    if not VARS['WRITE']:
+        return
     if not os.path.isfile(src):
         raise NotImplementedError(f'Source file not found: "{src}"')
     link(src, dst, desc)
@@ -29,6 +36,8 @@ def link_file(src, dst, desc=''):
 @decor_path_args(DOTFILES, HOME)
 def link_dir(src, dst, desc=''):
     '''Link file'''
+    if not VARS['WRITE']:
+        return
     if not os.path.isdir(src):
         raise NotImplementedError(f'Source directory not found: "{src}"')
     link(src, dst, desc, False)
@@ -40,7 +49,7 @@ def link(src, dst, desc, is_file=True):
     else:
         _type = 'file' if is_file else 'dir'
         desc = base_name(dst)
-        print_info(f'Replacing {_type} "{desc}":')
+        print_info(f'Replacing {_type}: "{desc}"')
 
     if os.path.isdir(dst):
         raise NotImplementedError(f'Destination is a directory: "{dst}"')
@@ -60,13 +69,15 @@ def read(dst):
 # Update
 def write(data, dst):
     '''Write data to file'''
+    if not VARS['WRITE']:
+        return
     with open(dst, 'w') as _f:
-        print_debug(f'Writing to "{dst}"')
+        print_debug(f'Writing to: "{dst}"')
         print_trace(f'Data: {data}')
 
 
 # Delete
 @decor_path
 def rm_file(path):
-    if os.path.isfile(path):
+    if VARS['WRITE'] and os.path.isfile(path):
         bash_cmd(f'rm "{path}"')
